// sdk/ai/ai-agents/src/models.ts
export interface ToolDefinitionParent {
  type: string;
}

/** The input definition information for a code interpreter tool as used to configure an agent. */
export interface CodeInterpreterToolDefinition extends ToolDefinitionParent {
  /** The object type, which is always 'code_interpreter'. */
  type: "code_interpreter";
}

/** The input definition information for a file search tool as used to configure an agent. */
export interface FileSearchToolDefinition extends ToolDefinitionParent {
  /** The object type, which is always 'file_search'. */
  type: "file_search";
  /** Options overrides for the file search tool. */
  file_search?: FileSearchToolDefinitionDetails;
}

/** Options overrides for the file search tool. */
export interface FileSearchToolDefinitionDetails {
  /**
   * The maximum number of results the file search tool should output. The default is 20 for gpt-4* models and 5 for gpt-3.5-turbo. This number should be between 1 and 50 inclusive.
   *
   * Note that the file search tool may output fewer than `max_num_results` results. See the file search tool documentation for more information.
   */
  max_num_results?: number;
  /** Ranking options for file search. */
  ranking_options?: FileSearchRankingOptions;
}

/** Ranking options for file search. */
export interface FileSearchRankingOptions {
  /** File search ranker. */
  ranker: string;
  /** Ranker search threshold. */
  score_threshold: number;
}

/** The input definition information for a function tool as used to configure an agent. */
export interface FunctionToolDefinition extends ToolDefinitionParent {
  /** The object type, which is always 'function'. */
  type: "function";
  /** The definition of the concrete function that the function tool should call. */
  function: FunctionDefinition;
}

/** The input definition information for a function. */
export interface FunctionDefinition {
  /** The name of the function to be called. */
  name: string;
  /** A description of what the function does, used by the model to choose when and how to call the function. */
  description?: string;
  /** The parameters the functions accepts, described as a JSON Schema object. */
  parameters: unknown;
}

/** The input definition information for a bing grounding search tool as used to configure an agent. */
export interface BingGroundingToolDefinition extends ToolDefinitionParent {
  /** The object type, which is always 'bing_grounding'. */
  type: "bing_grounding";
  /** The list of search configurations used by the bing grounding tool. */
  bing_grounding: BingGroundingSearchConfigurationList;
}

/** A list of search configurations currently used by the `bing_grounding` tool. */
export interface BingGroundingSearchConfigurationList {
  /**
   * The search configurations attached to this tool. There can be a maximum of 1
   * search configuration resource attached to the tool.
   */
  search_configurations: Array<BingGroundingSearchConfiguration>;
}

/** Search configuration for Bing Grounding */
export interface BingGroundingSearchConfiguration {
  /** Connection id for grounding with bing search */
  connection_id: string;
  /** The market where the results come from. */
  market?: string;
  /** The language to use for user interface strings when calling Bing API. */
  set_lang?: string;
  /** The number of search results to return in the bing api response */
  count?: number;
  /** Filter search results by a specific time range. Accepted values: https://learn.microsoft.com/bing/search-apis/bing-web-search/reference/query-parameters */
  freshness?: string;
}

/** The input definition information for a Microsoft Fabric tool as used to configure an agent. */
export interface MicrosoftFabricToolDefinition extends ToolDefinitionParent {
  /** The object type, which is always 'fabric_dataagent'. */
  type: "fabric_dataagent";
  /** The list of connections used by the Microsoft Fabric tool. */
  fabric_dataagent: ToolConnectionList;
}

/** A set of connection resources currently used by either the `bing_grounding`, `fabric_dataagent`, or `sharepoint_grounding` tools. */
export interface ToolConnectionList {
  /**
   * The connections attached to this tool. There can be a maximum of 1 connection
   * resource attached to the tool.
   */
  connections?: Array<ToolConnection>;
}

/** A connection resource. */
export interface ToolConnection {
  /** A connection in a ToolConnectionList attached to this tool. */
  connection_id: string;
}

/** The input definition information for a sharepoint tool as used to configure an agent. */
export interface SharepointToolDefinition extends ToolDefinitionParent {
  /** The object type, which is always 'sharepoint_grounding'. */
  type: "sharepoint_grounding";
  /** The list of connections used by the SharePoint tool. */
  sharepoint_grounding: ToolConnectionList;
}

/** The input definition information for an Azure AI search tool as used to configure an agent. */
export interface AzureAISearchToolDefinition extends ToolDefinitionParent {
  /** The object type, which is always 'azure_ai_search'. */
  type: "azure_ai_search";
}

/** The input definition information for an OpenAPI tool as used to configure an agent. */
export interface OpenApiToolDefinition extends ToolDefinitionParent {
  /** The object type, which is always 'openapi'. */
  type: "openapi";
  /** The openapi function definition. */
  openapi: OpenApiFunctionDefinition;
}

/** The input definition information for an openapi function. */
export interface OpenApiFunctionDefinition {
  /** The name of the function to be called. */
  name: string;
  /** A description of what the function does, used by the model to choose when and how to call the function. */
  description?: string;
  /** The openapi function shape, described as a JSON Schema object. */
  spec: unknown;
  /** Open API authentication details */
  auth: OpenApiAuthDetails;
  /** List of OpenAPI spec parameters that will use user-provided defaults */
  default_params?: string[];
}

/** authentication details for OpenApiFunctionDefinition */
export interface OpenApiAuthDetailsParent {
  type: OpenApiAuthType;
}

/** Security details for OpenApi anonymous authentication */
export interface OpenApiAnonymousAuthDetails extends OpenApiAuthDetailsParent {
  /** The object type, which is always 'anonymous'. */
  type: "anonymous";
}

/** Security details for OpenApi connection authentication */
export interface OpenApiConnectionAuthDetails extends OpenApiAuthDetailsParent {
  /** The object type, which is always 'connection'. */
  type: "connection";
  /** Connection auth security details */
  security_scheme: OpenApiConnectionSecurityScheme;
}

/** Security scheme for OpenApi managed_identity authentication */
export interface OpenApiConnectionSecurityScheme {
  /** Connection id for Connection auth type */
  connection_id: string;
}

/** Security details for OpenApi managed_identity authentication */
export interface OpenApiManagedAuthDetails extends OpenApiAuthDetailsParent {
  /** The object type, which is always 'managed_identity'. */
  type: "managed_identity";
  /** Connection auth security details */
  security_scheme: OpenApiManagedSecurityScheme;
}

/** Security scheme for OpenApi managed_identity authentication */
export interface OpenApiManagedSecurityScheme {
  /** Authentication scope for managed_identity auth type */
  audience: string;
}

/** The input definition information for a Bing custom search tool as used to configure an agent. */
export interface BingCustomSearchToolDefinition extends ToolDefinitionParent {
  /** The object type, which is always 'bing_custom_search'. */
  type: "bing_custom_search";
  /** The list of search configurations used by the bing custom search tool. */
  bing_custom_search: BingCustomSearchConfigurationList;
}

/** A list of search configurations currently used by the `bing_custom_search` tool. */
export interface BingCustomSearchConfigurationList {
  /**
   * The connections attached to this tool. There can be a maximum of 1 connection
   * resource attached to the tool.
   */
  search_configurations: Array<BingCustomSearchConfiguration>;
}
// ...
export interface ConnectedAgentToolDefinition extends ToolDefinitionParent {
  /** The object type, which is always 'connected_agent'. */
  type: "connected_agent";
  /** The sub-agent to connect */
  connected_agent: ConnectedAgentDetails;
}

/** Information for connecting one agent to another as a tool */
export interface ConnectedAgentDetails {
  /** The identifier of the child agent. */
  id: string;
  /** The name of the agent to be called. */
  name: string;
  /** A description of what the agent does, used by the model to choose when and how to call the agent. */
  description: string;
}

/** The input definition information for a azure function tool as used to configure an agent. */
export interface AzureFunctionToolDefinition extends ToolDefinitionParent {
  /** The object type, which is always 'azure_function'. */
  type: "azure_function";
  /** The definition of the concrete function that the function tool should call. */
  azure_function: AzureFunctionDefinition;
}

/** The definition of Azure function. */
export interface AzureFunctionDefinition {
  /** The definition of azure function and its parameters. */
  function: FunctionDefinition;
  /** Input storage queue. The queue storage trigger runs a function as messages are added to it. */
  input_binding: AzureFunctionBinding;
  /** Output storage queue. The function writes output to this queue when the input items are processed. */
  output_binding: AzureFunctionBinding;
}
// ...
export interface ToolResources {
  /** Resources to be used by the `code_interpreter` tool consisting of file IDs. */
  code_interpreter?: CodeInterpreterToolResource;
  /** Resources to be used by the `file_search` tool consisting of vector store IDs. */
  file_search?: FileSearchToolResource;
  /** Resources to be used by the `azure_ai_search` tool consisting of index IDs and names. */
  azure_ai_search?: AzureAISearchResource;
}

/** A set of resources that are used by the `code_interpreter` tool. */
export interface CodeInterpreterToolResource {
  /**
   * A list of file IDs made available to the `code_interpreter` tool. There can be a maximum of 20 files
   * associated with the tool.
   */
  file_ids?: string[];
  /** The data sources to be used. This option is mutually exclusive with the `fileIds` property. */
  data_sources?: Array<VectorStoreDataSource>;
}

/**
 * The structure, containing Azure asset URI path and the asset type of the file used as a data source
 * for the enterprise file search.
 */
export interface VectorStoreDataSource {
  /** Asset URI. */
  uri: string;
  /**
   * The asset type
   *
   * Possible values: "uri_asset", "id_asset"
   */
  type: VectorStoreDataSourceAssetType;
}

/** A set of resources that are used by the `file_search` tool. */
export interface FileSearchToolResource {
  /**
   * The ID of the vector store attached to this agent. There can be a maximum of 1 vector
   * store attached to the agent.
   */
  vector_store_ids?: string[];
  /**
   * The list of vector store configuration objects from Azure.
   * This list is limited to one element.
   * The only element of this list contains the list of azure asset IDs used by the search tool.
   */
  vector_stores?: Array<VectorStoreConfigurations>;
}

/** The structure, containing the list of vector storage configurations i.e. the list of azure asset IDs. */
export interface VectorStoreConfigurations {
  /** Name */
  name: string;
  /** Configurations */
  configuration: VectorStoreConfiguration;
}

/**
 * Vector storage configuration is the list of data sources, used when multiple
 * files can be used for the enterprise file search.
 */
export interface VectorStoreConfiguration {
  /** Data sources */
  data_sources: Array<VectorStoreDataSource>;
}

/** A set of index resources used by the `azure_ai_search` tool. */
export interface AzureAISearchResource {
  /**
   * The indices attached to this agent. There can be a maximum of 1 index
   * resource attached to the agent.
   */
  indexes?: Array<AISearchIndexResource>;
}

/** A AI Search Index resource. */
export interface AISearchIndexResource {
  /** An index connection id in an IndexResource attached to this agent. */
  index_connection_id: string;
  /** The name of an index in an IndexResource attached to this agent. */
  index_name: string;
  /**
   * Type of query in an AIIndexResource attached to this agent.
   *
   * Possible values: "simple", "semantic", "vector", "vector_simple_hybrid", "vector_semantic_hybrid"
   */
  query_type?: AzureAISearchQueryType;
  /** Number of documents to retrieve from search and present to the model. */
  top_k?: number;
  /** filter string for search resource. */
  filter?: string;
  /** Index asset id for search resource. */
  index_asset_id?: string;
}
// ...
export type ToolDefinition =
  | ToolDefinitionParent
  | CodeInterpreterToolDefinition
  | FileSearchToolDefinition
  | FunctionToolDefinition
  | BingGroundingToolDefinition
  | MicrosoftFabricToolDefinition
  | SharepointToolDefinition
  | AzureAISearchToolDefinition
  | OpenApiToolDefinition
  | BingCustomSearchToolDefinition
  | ConnectedAgentToolDefinition
  | AzureFunctionToolDefinition;
/** authentication details for OpenApiFunctionDefinition */
export type OpenApiAuthDetails =
  | OpenApiAuthDetailsParent
  | OpenApiAnonymousAuthDetails
  | OpenApiConnectionAuthDetails
  | OpenApiManagedAuthDetails;
// ...
export type VectorStoreDataSourceAssetType = string;
/** Alias for AzureAISearchQueryType */
export type AzureAISearchQueryType = string;